Implemented endpoint for pre-check report export and integrated with frontend
