+++ backend/src/controllers/prechecks-controller.ts
@@ -1,8 +1,7 @@
 import { Request, Response } from "express";
 import { getAllElasticNodes, getElasticNodeById } from "../services/elastic-node.service.";
-import { getLatestRunsByPrecheck, getMergedPrecheckStatus, runPrecheck } from "../services/precheck-runs.service";
-import { getPrecheckById } from "../config/precheck-config";
-import { INodePrecheckRun } from "../models/node-precheck-runs.model";
+import { getPrechecksGroupedByNode, runPrecheck } from "../services/precheck-runs.service";
+import { generatePrecheckReportMdContent } from "../services/precheck-report.service";
 
 export const runAllPrecheksHandler = async (req: Request, res: Response) => {
 	const { clusterId } = req.params;
@@ -24,45 +23,22 @@ export const runPrechekByNodeIdHandler = async (req: Request, res: Response) => 
 
 export const getPrecheckRunByClusterIdHandler = async (req: Request, res: Response) => {
 	const { clusterId } = req.params;
-	const precheckRuns = await getLatestRunsByPrecheck(clusterId);
-	if (!precheckRuns || precheckRuns.length === 0) {
-		res.status(404).send({ err: "No precheck runs found" });
-		return;
+	try {
+		const response = await getPrechecksGroupedByNode(clusterId);
+		res.send(response);
+	} catch (err: any) {
+		res.sendStatus(404).send(err.message);
 	}
-	const groupedPrecheckRunsByNodeId = precheckRuns.reduce<Record<string, typeof precheckRuns>>((acc, run) => {
-		const groupedBy = run.nodeId;
-		if (!acc[groupedBy]) {
-			acc[groupedBy] = [];
-		}
-		acc[groupedBy].push(run);
-		return acc;
-	}, {});
+};
 
-	const response = Object.entries(groupedPrecheckRunsByNodeId).map(([nodeId, precheckRuns]) => {
-		const status = getMergedPrecheckStatus(precheckRuns.map((precheck) => precheck.status));
-		const precheck = precheckRuns[0];
-		const transformPrecheckRunForUI = (precheck: INodePrecheckRun) => {
-			const { name } = getPrecheckById(precheck.precheckId) || {};
-			const duration = precheck.endAt
-				? parseFloat(((precheck.endAt.getTime() - precheck.startedAt.getTime()) / 1000).toFixed(2))
-				: null;
-			return {
-				id: precheck.precheckId,
-				name: name,
-				status: precheck.status,
-				logs: precheck.logs,
-				startTime: precheck.startedAt,
-				endTime: precheck.endAt,
-				duration: duration,
-			};
-		};
-		return {
-			nodeId: nodeId,
-			ip: precheck.ip,
-			name: precheck.nodeName,
-			status: status,
-			prechecks: precheckRuns.map(transformPrecheckRunForUI),
-		};
-	});
-	res.send(response);
+export const getPrecheckReportByClusterId = async (req: Request<{ clusterId: string }>, res: Response) => {
+	const { clusterId } = req.params;
+	try {
+		const content = await generatePrecheckReportMdContent(clusterId);
+		res.setHeader("Content-Disposition", `attachment; filename="precheck-report.md"`);
+		res.setHeader("Content-Type", "text/markdown");
+		res.send(content);
+	} catch (err: any) {
+		res.sendStatus(404).send(err.message);
+	}
 };
