+++ backend/src/services/precheck-report.service.ts
@@ -1,0 +1,8 @@
+import { getPrechecksGroupedByNode } from "./precheck-runs.service";
+
+export const generatePrecheckReportMdContent = async (clusterId: string): Promise<string> => {
+	const prechecksGroupedByNode = await getPrechecksGroupedByNode(clusterId);
+	return `
+  # Precheck Report 
+  `;
+};
